Moved RelationNameRepository a package level up
