+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/Relation.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.*;
+import static com.dat3m.dartagnan.wmm.RelationNameRepository.*;
 import static com.google.common.base.Preconditions.checkState;
 
 public final class Relation implements Dependent<Relation> {
