+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/definition/DifferentThreads.java
@@ -3,7 +3,7 @@ package com.dat3m.dartagnan.wmm.definition;
 import com.dat3m.dartagnan.wmm.Definition;
 import com.dat3m.dartagnan.wmm.Relation;
 
-import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.EXT;
+import static com.dat3m.dartagnan.wmm.RelationNameRepository.EXT;
 
 public class DifferentThreads extends Definition {
