+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/definition/DirectControlDependency.java
@@ -3,7 +3,7 @@ package com.dat3m.dartagnan.wmm.definition;
 import com.dat3m.dartagnan.wmm.Definition;
 import com.dat3m.dartagnan.wmm.Relation;
 
-import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.CTRLDIRECT;
+import static com.dat3m.dartagnan.wmm.RelationNameRepository.CTRLDIRECT;
 
 //TODO: We can restrict the codomain to visible events as the only usage of this Relation is in
 // ctrl := idd^+;ctrlDirect & (R*V)
