+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/definition/Empty.java
@@ -2,7 +2,7 @@ package com.dat3m.dartagnan.wmm.definition;
 
 import com.dat3m.dartagnan.wmm.Definition;
 import com.dat3m.dartagnan.wmm.Relation;
-import com.dat3m.dartagnan.wmm.relation.RelationNameRepository;
+import com.dat3m.dartagnan.wmm.RelationNameRepository;
 
 public class Empty extends Definition {
