+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/definition/SameThread.java
@@ -3,7 +3,7 @@ package com.dat3m.dartagnan.wmm.definition;
 import com.dat3m.dartagnan.wmm.Definition;
 import com.dat3m.dartagnan.wmm.Relation;
 
-import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.INT;
+import static com.dat3m.dartagnan.wmm.RelationNameRepository.INT;
 
 public class SameThread extends Definition {
